dummy1: fix branch and add copy_files example
This sample target uses a git that track master instead of main. Update
to main to avoid getting checkout errors when setting up the workspace.

Add an "extra.mk" that we both refer to as a makefile include as well as
a copy_files to get included in the workspace.

Signed-off-by: Joakim Bech <Joakim.Bech@analog.com>