no-OS: Add swiot1l as a first test target
Add a new target 'swiot1l', which perhaps will be renamed in the future
to no-OS or some other more suitable name. This target will clone no-OS
and msdk and download the necessary toolchains (only verified on Linux
aarch64 so far).

We also clone openocd, but that target hasn't been configured, so
instead we get openocd from the distro packages.

Steps to build this setup is:
$ sdk-manager init -t swiot1l
$ cd  dsdk-workspace
$ sdk-manager install os-deps
$ sdk-manager install toolchains
$ sdk-manager makefile
$ make sdk-envsetup
$ make sdk-build

Note, we can only run "make sdk-envsetup", since we apply patches when
doing that and currently we have no check whether they already have been
applied or not. Ideally we shouldn't do patching at all, that's just a
quick workaround for now.

Signed-off-by: Joakim Bech <Joakim.Bech@analog.com>